Added safety checks and fixed CMP zero flag behaviour.
The patches are basically not to read past end of RAM or ROM, should a ROM instruction try to do so.
Additionally, CMP is no longer "sticky" i.e. the zero flag is set to true if CMP is true (as before), but now the zero flag is set to false if the CMP is false. As a consequence, the JZ/JNZ instructions should now work as intended.